fix: separate type in result in query file
