+++ app/business/user/user.type.ts
@@ -1,7 +1,7 @@
 // https://stackoverflow.com/questions/76957592/error-only-async-functions-are-allowed-to-be-exported-in-a-use-server-file
 // server action 파일에서는 async function만 export 가능
 
-import { Major } from '../result/result.query';
+import { Major } from '../result/result.type';
 import { SignInResponseSchema, UserInfoResponseSchema, ValidateTokenResponseSchema } from './user.validation';
 import z from 'zod';
